+++ packages/bundlekit-cli/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @bundlekit/cli
 
+## 0.0.17
+
+### Patch Changes
+
+- 8e12ab8: 修复项目初始化异常和启动开发环境异常
+- Updated dependencies [8e12ab8]
+  - @bundlekit/shared-utils@0.0.9
+  - @bundlekit/plugin-node@0.0.4
+  - @bundlekit/plugin-react@0.0.12
+  - @bundlekit/plugin-vue@0.0.12
+
 ## 0.0.16
 
 ### Patch Changes